[Chore] #534 - StringLiterals extension을 활용한 기능 별 파일 분리
